requirements.txt
# Core ML/DL dependencies
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Data Processing
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0

# Configuration & Utilities
PyYAML>=6.0
tqdm>=4.62.0
click>=8.0.0
python-dotenv>=0.19.0

# Development & Testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Optional: For advanced deployment
# onnx>=1.12.0
# onnxruntime>=1.12.0
# tensorrt>=8.4.0  # For NVIDIA deployment
# coremltools>=5.0.0  # For Apple deployment